ignition defaults
