+++ firmware/controllers/algo/advance_map.h
@@ -10,7 +10,6 @@
 #include "engine.h"
 
 angle_t getAdvance(int rpm, float engineLoad DECLARE_ENGINE_PARAMETER_SUFFIX);
-void setDefaultIatTimingCorrection(DECLARE_ENGINE_PARAMETER_SIGNATURE);
 void initTimingMap(DECLARE_ENGINE_PARAMETER_SIGNATURE);
 float getTopAdvanceForBore(chamber_style_e style, int octane, double compression, double bore);
 float getInitialAdvance(int rpm, float map, float advanceMax);
