+++ firmware/controllers/algo/algo.mk
@@ -23,3 +23,4 @@ CONTROLLERS_ALGO_SRC_CPP = $(PROJECT_DIR)/controllers/algo/advance_map.cpp \
 	$(PROJECT_DIR)/controllers/algo/nmea.cpp \
 	$(PROJECT_DIR)/controllers/algo/defaults/default_base_engine.cpp \
 	$(PROJECT_DIR)/controllers/algo/defaults/default_fuel.cpp \
+	$(PROJECT_DIR)/controllers/algo/defaults/default_ignition.cpp \
