+++ firmware/controllers/algo/defaults/defaults.h
@@ -2,3 +2,4 @@
 
 void setDefaultBaseEngine(DECLARE_CONFIG_PARAMETER_SIGNATURE);
 void setDefaultFuel(DECLARE_CONFIG_PARAMETER_SIGNATURE);
+void setDefaultIgnition(DECLARE_CONFIG_PARAMETER_SIGNATURE);
